Add page for viewing slow/error queries to admin area. Fix CSRF vulnerability in admin/pageDeleteGroup.php, admin/pageDeleteMember.php and admin/pageDeleteRecord.php

+++ app/admin/pageDeleteMember.php
@@ -5,6 +5,8 @@
 	// validate input
 	$memberID=makeSafe(strtolower($_GET['memberID']));
 
+	if(!csrf_token(true)) die($Translation['csrf token expired or invalid']);
+
 	sql("delete from membership_users where lcase(memberID)='$memberID'", $eo);
 	sql("update membership_userrecords set memberID='' where lcase(memberID)='$memberID'", $eo);
