+++ app/admin/pageDeleteRecord.php
@@ -5,6 +5,8 @@
 	// validate input
 	$recID = intval($_GET['recID']);
 
+	if(!csrf_token(true)) die($Translation['csrf token expired or invalid']);
+
 	$res = sql("SELECT `tableName`, `pkValue` FROM `membership_userrecords` WHERE `recID`='{$recID}'", $eo);
 	if($row = db_fetch_row($res)) {
 		sql("DELETE FROM `membership_userrecords` WHERE `recID`='{$recID}'", $eo);
